add texture animation

+++ ParticleSystem3D/Renderer/PS3RendererCPU.hpp
@@ -27,7 +27,7 @@ public:
     void Render() override; // 在原版里是beforeRender
     
 public:
-    void FillMeshData(PS3ParticlePtr particle, int idx, int fi); // 本来应该根据RenderMode有多种func，暂时只写FillMeshData
+    void FillMeshData(PS3ParticlePtr particle, int idx, float fi); // 本来应该根据RenderMode有多种func，暂时只写FillMeshData
     
 public:
     std::vector<PS3ParticlePtr> _particles; // 粒子池
